Adds caching of exclusion warnings
The authentication exclusion warnings can be very chatty and distracting
on quiet services, especially if automated monitoring tools are hitting
an excluded endpoint e.g. /healthz or equivalent.  This commit adds in
an exclusion warning cache that is used to rate limit how often users
are warned about path exclusions to reduce the frequency of these
warnings appearing in the logs.  Cache is set to small size (10) and
reasonable duration (15 minutes) so badly configured services with too
many excluded paths will still see plenty of warnings in their logs.