Add explicit error_message to initial 401 Challenge
In some scenarios, typically misconfigured auth, or clients not
pre-emptively sending auth credentials then a 401 Challenge is generated
with a blank message.  While this is perfectly correct from a HTTP
perspective, where applications were logging the challenge error_message
as part of application level logging this led to weird blank error
messages in the logs.

This commit adds a default error_message for this scenario indicating
what is happening, and makes it possible for derived engines to provide
a more specific error_message if desired.